feat: add subsection models and DTOs for sightseeing, activity, and route
