+++ src/main/java/com/devs/roamance/dto/request/travel/journal/ActivitySubsectionCreateRequestDto.java
@@ -1,11 +1,19 @@
-package com.devs.roamance.dto.request.subsection;
+package com.devs.roamance.dto.request.travel.journal;
 
 import com.devs.roamance.dto.request.LocationCreateRequestDto;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Size;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
 public class ActivitySubsectionCreateRequestDto extends SubsectionCreateRequestDto {
     @Valid
     @NotNull(message = "Location is required for activity")
