+++ src/main/java/com/devs/roamance/dto/request/travel/journal/ActivitySubsectionUpdateRequestDto.java
@@ -1,9 +1,17 @@
-package com.devs.roamance.dto.request.subsection;
+package com.devs.roamance.dto.request.travel.journal;
 
 import com.devs.roamance.dto.request.LocationUpdateRequestDto;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Size;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
 public class ActivitySubsectionUpdateRequestDto extends SubsectionUpdateRequestDto {
     @Valid
     private LocationUpdateRequestDto location;
