+++ src/main/java/com/devs/roamance/dto/request/travel/journal/JournalCreateRequestDto.java
@@ -1,6 +1,6 @@
-package com.devs.roamance.dto.request;
+package com.devs.roamance.dto.request.travel.journal;
 
-import com.devs.roamance.dto.request.subsection.SubsectionCreateRequestDto;
+import com.devs.roamance.dto.request.LocationCreateRequestDto;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -21,9 +21,11 @@ public class JournalCreateRequestDto {
     @NotBlank(message = "Title is required")
     @Size(max = 100, message = "Title must be less than 100 characters")
     private String title;
+
     @Valid
     @NotNull(message = "Destination is required")
     private LocationCreateRequestDto destination;
+
     @Size(max = 1000, message = "Description must be less than 1000 characters")
     private String description;
