+++ src/main/java/com/devs/roamance/dto/request/travel/journal/SightseeingSubsectionCreateRequestDto.java
@@ -1,9 +1,17 @@
-package com.devs.roamance.dto.request.subsection;
+package com.devs.roamance.dto.request.travel.journal;
 
 import com.devs.roamance.dto.request.LocationCreateRequestDto;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
 public class SightseeingSubsectionCreateRequestDto {
     @Valid
     @NotNull(message = "Location is required for sightseeing")
