+++ src/main/java/com/devs/roamance/dto/request/travel/journal/SightseeingSubsectionUpdateRequestDto.java
@@ -1,12 +1,16 @@
-package com.devs.roamance.dto.request.subsection;
+package com.devs.roamance.dto.request.travel.journal;
 
 import com.devs.roamance.dto.request.LocationUpdateRequestDto;
 import jakarta.validation.Valid;
+import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 @Getter
 @Setter
+@NoArgsConstructor
+@AllArgsConstructor
 public class SightseeingSubsectionUpdateRequestDto extends SubsectionUpdateRequestDto {
     @Valid
     private LocationUpdateRequestDto location;
