+++ src/main/java/com/devs/roamance/dto/request/travel/journal/SubsectionCreateRequestDto.java
@@ -1,8 +1,9 @@
-package com.devs.roamance.dto.request.subsection;
+package com.devs.roamance.dto.request.travel.journal;
 
-import com.devs.roamance.model.subsection.SubsectionType;
+import com.devs.roamance.model.travel.journal.SubsectionType;
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
+import jakarta.validation.Valid;
 import jakarta.validation.constraints.*;
 import lombok.Getter;
 import lombok.Setter;
@@ -22,7 +23,7 @@ public abstract class SubsectionCreateRequestDto {
     @Size(max = 100, message = "Title must be less than 100 characters")
     private String title;
 
-    @NotNull(message = "Subsection type is required")
+    @Valid
     private SubsectionType type;
 
     @Size(max = 10, message = "Maximum 10 notes allowed")
