--- src/main/java/com/devs/roamance/dto/response/JournalBriefResponseDto.java
@@ -1,21 +1,0 @@
-package com.devs.roamance.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.time.ZonedDateTime;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class JournalBriefResponseDto {
-    private Long id;
-    private String title;
-    private LocationResponseDto destination;
-    private String description;
-    private int subsectionCount;
-    private ZonedDateTime lastModified;
-}
