+++ src/main/java/com/devs/roamance/dto/response/travel/journal/JournalBriefResponseDto.java
@@ -1,22 +1,22 @@
-package com.devs.roamance.dto.response;
+package com.devs.roamance.dto.response.travel.journal;
 
+import com.devs.roamance.dto.response.LocationResponseDto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+
 import java.time.ZonedDateTime;
 
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class JournalResponseDto {
+public class JournalBriefResponseDto {
     private Long id;
     private String title;
     private LocationResponseDto destination;
     private String description;
-    private ZonedDateTime createdAt;
+    private int subsectionCount;
     private ZonedDateTime lastModified;
-    private long createdBy;
-    private long lastModifiedBy;
 }
