+++ src/main/java/com/devs/roamance/dto/response/travel/journal/SubsectionResponseDto.java
@@ -1,6 +1,6 @@
-package com.devs.roamance.dto.response.subsection;
+package com.devs.roamance.dto.response.travel.journal;
 
-import com.devs.roamance.model.subsection.SubsectionType;
+import com.devs.roamance.model.travel.journal.SubsectionType;
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import lombok.AllArgsConstructor;
